+++ src/circuit_simulator/__init__.py
@@ -1,0 +1,7 @@
+from .Element import Element
+from .Circuit import Circuit
+from .Simulation import Simulation
+from .SimulationConfig import SimulationConfig
+
+# Agora todos os elementos podem ser importados diretamente do pacote
+__all__ = ["Element", "Circuit", "Simulation", "SimulationConfig"]
