+++ src/circuit_simulator/elements/__init__.py
@@ -1,0 +1,8 @@
+from .Capacitor import Capacitor
+from .FTCT import FTCT
+from .Indutor import Indutor
+from .Resistor import Resistor
+from .ResistorNL import ResistorNL
+
+# Agora todos os elementos podem ser importados diretamente do pacote
+__all__ = ["Resistor", "Indutor", "Capacitor", "ResistorNL", "FTCT"]
